refactor album iterator + add fn to get user owned playlsits
